+++ frontend/app/src/components/defi/dex-trades/DexTradeHistory.vue
@@ -1,6 +1,10 @@
 <template>
   <v-container>
-    <no-premium-placeholder v-if="!premium" :text="$t('dex_trade.title')" />
+    <module-not-active v-if="!isUniswapEnabled" :module="MODULE_UNISWAP" />
+    <no-premium-placeholder
+      v-else-if="!premium"
+      :text="$t('dex_trade.title')"
+    />
     <progress-screen v-else-if="loading">
       <template #message>{{ $t('dex_trades.loading') }}</template>
     </progress-screen>
@@ -11,20 +15,31 @@
 <script lang="ts">
 import { Component, Mixins } from 'vue-property-decorator';
 import { mapActions } from 'vuex';
+import ModuleNotActive from '@/components/defi/ModuleNotActive.vue';
 import ProgressScreen from '@/components/helper/ProgressScreen.vue';
 import NoPremiumPlaceholder from '@/components/premium/NoPremiumPlaceholder.vue';
+import DefiModuleMixin from '@/mixins/defi-module-mixin';
 import PremiumMixin from '@/mixins/premium-mixin';
 import StatusMixin from '@/mixins/status-mixin';
 import { Section } from '@/store/const';
 import { DexTradesTable } from '@/utils/premium';
 
 @Component({
-  components: { NoPremiumPlaceholder, ProgressScreen, DexTradesTable },
+  components: {
+    ModuleNotActive,
+    NoPremiumPlaceholder,
+    ProgressScreen,
+    DexTradesTable
+  },
   methods: {
     ...mapActions('defi', ['fetchUniswapTrades'])
   }
 })
-export default class DexTradeHistory extends Mixins(StatusMixin, PremiumMixin) {
+export default class DexTradeHistory extends Mixins(
+  StatusMixin,
+  PremiumMixin,
+  DefiModuleMixin
+) {
   section = Section.DEFI_UNISWAP_TRADES;
   fetchUniswapTrades!: (refresh: boolean) => Promise<void>;