+++ frontend/app/src/components/defi/uniswap/Uniswap.vue
@@ -1,7 +1,16 @@
 <template>
-  <progress-screen v-if="loading">
+  <module-not-active v-if="!isUniswapEnabled" :module="MODULE_UNISWAP" />
+  <progress-screen v-else-if="loading">
     <template #message>
       {{ $t('uniswap.loading') }}
+    </template>
+    <template v-if="!premium" #default>
+      <i18n tag="div" path="uniswap.loading_non_premium">
+        <base-external-link
+          :text="$t('uniswap.premium')"
+          :href="$interop.premiumURL"
+        />
+      </i18n>
     </template>
   </progress-screen>
   <v-container v-else class="uniswap">
@@ -157,9 +166,13 @@
 <script lang="ts">
 import { Component, Mixins } from 'vue-property-decorator';
 import { mapActions, mapGetters } from 'vuex';
+import BaseExternalLink from '@/components/base/BaseExternalLink.vue';
+import ModuleNotActive from '@/components/defi/ModuleNotActive.vue';
 import UniswapPoolAsset from '@/components/display/icons/UniswapPoolAsset.vue';
 import BlockchainAccountSelector from '@/components/helper/BlockchainAccountSelector.vue';
 import ProgressScreen from '@/components/helper/ProgressScreen.vue';
+import DefiModuleMixin from '@/mixins/defi-module-mixin';
+import PremiumMixin from '@/mixins/premium-mixin';
 import StatusMixin from '@/mixins/status-mixin';
 import { UnknownToken } from '@/services/defi/types';
 import { SupportedAsset } from '@/services/types-model';
@@ -171,6 +184,8 @@ import { UniswapDetails } from '@/utils/premium';
 
 @Component({
   components: {
+    BaseExternalLink,
+    ModuleNotActive,
     UniswapDetails,
     ProgressScreen,
     UniswapPoolAsset,
@@ -184,7 +199,11 @@ import { UniswapDetails } from '@/utils/premium';
     ...mapActions('defi', ['fetchUniswapEvents', 'fetchUniswapBalances'])
   }
 })
-export default class Uniswap extends Mixins(StatusMixin) {
+export default class Uniswap extends Mixins(
+  StatusMixin,
+  DefiModuleMixin,
+  PremiumMixin
+) {
   readonly ETH = ETH;
   readonly assetName = assetName;
   section = Section.DEFI_UNISWAP_BALANCES;